requirements.txt
# Core dependencies
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
streamlit
python-dotenv
crawl4ai
beautifulsoup4
lxml
pandas  
tqdm
requests    
PyYAML


